Rust builds & most used Pathname#chop_basename
Rust will now build a native linked library extension.

Pathname#chop_basename has a bug in cases where blank strings are
provided.  The re-implementation does not have this problem so that is
the only changed behavior for this method.

FasterPath.absolute? behavior has not yet been compared to
Pathname#.absolute?