Improved speed of FasterPath.chop_basename again
An additional 19.2% performance gain has been added to
`FasterPath.chop_basename`.  This is the most called method in Rails
apps that I've seen.  With it's previous improvement of 27.5% my Rails
application performance improved by 15%.  So this is a huge performance
gain!

I've added more iterations to the benchmark test to try and get the
numbers tighter for more accurate benchmarks.

closes #5