New basename implementation 31.3% faster than Ruby
Had to write it from scratch with no path helpers or regex.
Performance has been improved.  This was my 3rd attempt for better
performance.  First I tried implementing regex chopping in Ruby and
that was 60X slower.  Then I tried regex in Rust and that was 5600X
slower.  So I started from scratch and use minimal types to implement
it and gained 31.3%.  Say 31.3% percent one more time I dare you ;-)

Fixes #7

Oh, you will see a copy of the method with a bunch of tests written in
Rust (for that unused method).  It's much faster to test this way.  Then
instead of using that method - copy the code over; since some raw types
can't be returned from a method in Rust (like str) and that will
cost performance.